requirements.txt
PyQt5
icecream
attrs
